refactor: removing local EntityAccess.java and implementing the one from evm-module external lib
