+++ hedera-mirror-web3/src/main/java/com/hedera/mirror/web3/repository/EntityAccessRepository.java
@@ -8,19 +8,13 @@ import com.hedera.mirror.common.domain.entity.Entity;
 import com.hedera.mirror.common.domain.entity.EntityType;
 
 public interface EntityAccessRepository extends CrudRepository<Entity, Long> {
-    @Query(value = "select balance from entity where realm = ?1 and shard = ?2 and num = ?3 and deleted <> true",
+    @Query(value = "select balance from entity where num = ?1 and deleted <> true",
             nativeQuery = true)
-    Optional<Long> getBalance(final Long realmNum, final Long shardNum,
-                              final Long accountNum);
+    Optional<Long> getBalance(final Long accountNum);
 
-    @Query(value = "select deleted  from entity where realm = ?1 and shard = ?2 and num = ?3", nativeQuery = true)
-    Optional<Boolean> isDeleted(final Long realmNum, final Long shardNum,
-                                final Long accountNum);
-
-    @Query(value = "select type from entity where realm = ?1 and shard = ?2 and num = ?3 and deleted <> true"
+    @Query(value = "select type from entity where num = ?1 and deleted <> true"
             , nativeQuery = true)
-    Optional<EntityType> getType(final Long realmNum, final Long shardNum,
-                                 final Long accountNum);
+    Optional<EntityType> getType(final Long accountNum);
 
     @Query(value = "select runtime_bytecode from contract where id = ?1",
             nativeQuery = true)
@@ -29,5 +23,4 @@ public interface EntityAccessRepository extends CrudRepository<Entity, Long> {
     @Query(value = "select value_written from contract_state_change where contract_id = ?1 and slot =?2",
             nativeQuery = true)
     byte[] getStorage(final Long accountNum, final byte[] key);
-
 }
